requirements.txt
streamlit>=1.32
spacy>=3.7
spacy-transformers>=1.3
langchain>=0.2
langgraph>=0.1
openai>=1.30
pymupdf>=1.24
python-docx>=1.1
faker>=24.0
pandas>=2.0
scikit-learn>=1.4
torch>=2.2
